Factor task management into the shared pacakge
This is definitely valid for at least DS3 forward, probably Bloodborne
as well. The two key extensibility points are the underlying function
to register a new task and the argument that's passed to the
task (which might be similar across games, but this isn't something
I've confirmed).

This does assume that the subset of the actual task interface that we
implement for out custom tasks is the same across games, but in
practice it does seem to work for both ER and DS3.