[ER] Move RVAs from eldenring-util into eldenring
The main value of making these usable from the base package is to make
it possible to implement traits like `SharedTaskImp` that use RVAs.
These traits can't be implemented outside the package that declares
the original struct, so anything they transitively touch has to be in
the same package (or one that's upstream of it).

This does mean that the RVA infrastructure is no longer
package-private, but in some cases this is useful—RVAs can be a handy
way to share things like function addresses for hooking even if they
don't necessarily make sense to provide util wrappers for.