[Shared] Add a MaybeEmpty struct
This allows us to represent structures that can either be initialized
or in an uninititialized *but known* state. Unlike `MaybeUninit`, the
backing memory is always set to some known pattern (often but not
always all zeroes), which we can use to provide a consistent, safe
interface for determining when and whether a struct is empty.